CRAS: Include cras_stream_apm.h in cras_stream_apm_noop.c
Code Search cannot find the definition of functions in
cras_stream_apm.h. It only finds the stubs defined in
unit tests.

When generate kythe indexes, we ./configure without webrtc-apm.
So it is expected that functions in cras_stream_apm.c cannot
be found. However it should still find the ones in
cras_stream_apm_noop.c.
I suspect that Kythe think that the functions in cras_stream_apm_noop.c
are different from the ones declared in cras_stream_apm.h,
probably due to the fact that cras_stream_apm_noop.c does
not #include "cras_stream_apm.h". For example, I observed that
in Code Search clicking on struct cras_stream_apm *stream
in cras_stream_apm_noop.c does not jump to the declaration in cras_stream_apm.h.

Let's get the header included and see what happens.

BUG=None
TEST=make check

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3697567
Tested-by: Li-Yu Yu <aaronyu@google.com>
Auto-Submit: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>

+++ cras/src/server/cras_stream_apm_noop.c
@@ -6,6 +6,7 @@
 #include <stdlib.h>
 
 #include "cras_iodev.h"
+#include "cras_stream_apm.h"
 #include "float_buffer.h"
 
 /*
@@ -24,7 +25,7 @@ int cras_stream_apm_deinit()
 {
 	return 0;
 }
-struct cras_stream_apm *cras_stream_apm_create(unsigned int effects)
+struct cras_stream_apm *cras_stream_apm_create(uint64_t effects)
 {
 	return NULL;
 }
